requirements.txt
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cramjam==2.6.1
cycler==0.11.0
dask==2022.10.2
distributed==2022.10.2
fonttools==4.38.0
fsspec==2022.10.0
greenlet==2.0.1
HeapDict==1.0.1
Jinja2==3.1.2
kiwisolver==1.4.4
lark-parser==0.12.0
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
msgpack==1.0.4
mypy==0.982
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pandas-stubs==1.5.1.221024
partd==1.3.0
Pillow==9.3.0
psutil==5.9.4
pyarrow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.43
tblib==1.7.0
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
tornado==6.1
tqdm==4.64.1
types-pytz==2022.6.0.1
typing-extensions==4.4.0
urllib3==1.26.12
zict==2.2.0
